refactor: extract sub-components to flatten nested JSX
Extract deeply nested JSX blocks into named components to improve
readability without changing any logic or styles.

- UnitRowExpanded from UnitRow.tsx (expanded section)
- UnitCardDetail from ExpandableUnitCard.tsx (IIFE elimination)
- UnitsTab, ShoppingTab from ArmyViewPage.tsx
- UnitsTab, StrategemsTab from FactionDetailPage.tsx
